+++ internal/distributed/proxy/httpserver/utils.go
@@ -25,7 +25,7 @@ import (
 	"github.com/milvus-io/milvus/pkg/util"
 	"github.com/milvus-io/milvus/pkg/util/funcutil"
 	"github.com/milvus-io/milvus/pkg/util/merr"
-	"github.com/milvus-io/milvus/pkg/util/parameterutil.go"
+	"github.com/milvus-io/milvus/pkg/util/parameterutil"
 )
 
 func ParseUsernamePassword(c *gin.Context) (string, string, bool) {
